Create ConnectionHandler for Server - Create ConnectionHandler.cpp - Create ConnectionHandler.h - Move socket connection logic to ConnectionHandler
